// src/enums/fusion.ts
export enum ScheduleType {
  Indefinitely = "indefinitely",
  Once = "once",
  Daily = "daily",
  Weekly = "weekly",
  Monthly = "monthly",
  Yearly = "yearly",
}

export enum SessionStatus {
  Building = "Building",
  Paused = "Paused",
  Complete = "Complete",
  UserCancelled = "UserCancel",
}

export enum EdgeConditionOperators {
  Exists = "exist",
  NotExists = "notexist",
  TextEqual = "text:equal",
  TextEqualCI = "text:equal:ci",
  TextNotEqual = "text:notequal",
  TextNotEqualCI = "text:notequal:ci",
  TextContains = "text:contain",
  TextContainsCI = "text:contain:ci",
  TextNotContains = "text:notcontain",
  TextNotContainsCI = "text:notcontain:ci",
  TextStartWith = "text:startwith",
  TextStartWithCI = "text:startwith:ci",
  TextNotStartWith = "text:notstartwith",
  TextNotStartWithCI = "text:notstartwith:ci",
  TextEndWith = "text:endwith",
  TextEndWithCI = "text:endwith:ci",
  TextNotEndWith = "text:notendwith",
  TextNotEndWithCI = "text:notendwith:ci",
  TextPattern = "text:pattern",
  TextPatternCI = "text:pattern:ci",
  TextNotPattern = "text:notpattern",
  TextNotPatternCI = "text:notpattern:ci",
  NumberEqual = "number:equal",
  NumberNotEqual = "number:notequal",
  NumberGreaterThan = "number:greater",
  NumberLessThan = "number:less",
  NumberGreaterThanOrEqual = "number:greaterorequal",
  NumberLessThanOrEqual = "number:lessorequal",
  DateEqual = "date:equal",
  DateNotEqual = "date:notequal",
  DateGreaterThan = "date:greater",
  DateLessThan = "date:less",
  DateGreaterThanOrEqual = "date:greaterorequal",
  DateLessThanOrEqual = "date:lessorequal",
  TimeEqual = "time:equal",
  TimeNotEqual = "time:notequal",
  TimeGreaterThan = "time:greater",
  TimeLessThan = "time:less",
  TimeGreaterThanOrEqual = "time:greaterorequal",
  TimeLessThanOrEqual = "time:lessorequal",
  BooleanEqual = "boolean:equal",
  BooleanNotEqual = "boolean:notequal",
  ArrayContains = "array:contain",
  ArrayContainsCI = "array:contain:ci",
  ArrayNotContains = "array:notcontain",
  ArrayNotContainsCI = "array:notcontain:ci",
  ArrayEquals = "array:equal",
  ArrayNotEquals = "array:notequal",
  ArrayGreater = "array:greater",
  ArrayLess = "array:less",
  ArrayGreaterOrEqual = "array:greaterorequal",
  ArrayLessOrEqual = "array:lessorequal",
}

export enum FusionType {
  Core = "core",
  Skills = "skills",
  Import = "import",
  Adventure = "adventure",
  DoubleTalk = "double_talk",
  SalineChat = "saline_chat",
  DataListWidgetCreateActionFormSubmit = "data-list-widget-create-action-form-submit",
  DataListWidgetEditActionFormSubmit = "data-list-widget-edit-action-form-submit",
  DataListWidgetCreateActionFormPopulate = "data-list-widget-create-action-form-populate",
  DataListWidgetEditActionFormPopulate = "data-list-widget-edit-action-form-populate",
  DataListGuiCreateFormSubmit = "data-list-gui-create-form-submit",
  DataList = "data-list",
  ReviewerAction = "fusion_action",
}